batch: Separate replacement selection from ownership assembly
The hunk translator projects selected displayed replacements into saved
presence and absence ownership.

I need to inspect selected-run policy without tracing origin projection and
ownership accumulation inside the same long function.

This commit separates lazy selected-run generation, a forward origin
cursor, and ownership assembly. The coordinator retains iterator lifetime
control and borrows existing line sequences.
